main: add binstub code to get the kernel compiling
This commit adds a stub multiboot application that purely writes garbage
to the screen by writing into the 0xB8000 memory address.  The bootMain
symbol runs a loop to write stuff into the memory address of the
standard textmode VGA framebuffer.

A Makefile script is also added to organize the compiling phases. The
following additional phony targets are added:

* qemu: use this to start qemu-system-i386 to run the kernel.
* clean: use this to clean all the generated artifacts.