console: write characters to framebuffer
This commit adds a unit called console with a few methods designed to
print into the framebuffer through the consolePutChar and
consolePutString procedures.

Additionally, consoleSetAttributes can be used to control the foreground
and background color; and consoleClearString can be used to clean and
reset the framebuffer.

When this commit is applied on top of the current Git head, the bootable
application no more writes garbage to the screen, but rather your
typical Hello world message.