main: set up a stack
This commit adds a 4 KB stack into the multiboot application.  This has
been done because I'm calling a lot of routines inside kernel.pas and
I'm a little concerned about the usage of the stack.  I defered
initially the implementation of a stack until I had the kernel booting,
but now it boots and so many subroutines request a stack to be added.